Adding backing vector3 to YCbCr.

+++ src/ImageSharp/Colors/ColorspaceTransforms.cs
@@ -63,13 +63,13 @@ namespace ImageSharp
         /// </returns>
         public static implicit operator Color(YCbCr color)
         {
-            byte y = color.Y;
-            int cb = color.Cb - 128;
-            int cr = color.Cr - 128;
+            float y = color.Y;
+            float cb = color.Cb - 128;
+            float cr = color.Cr - 128;
 
-            byte r = (byte)(y + (1.402F * cr)).Clamp(0, 255);
-            byte g = (byte)(y - (0.34414F * cb) - (0.71414F * cr)).Clamp(0, 255);
-            byte b = (byte)(y + (1.772F * cb)).Clamp(0, 255);
+            float r = (y + (1.402F * cr)).Clamp(0, 255);
+            float g = (y - (0.34414F * cb) - (0.71414F * cr)).Clamp(0, 255);
+            float b = (y + (1.772F * cb)).Clamp(0, 255);
 
             return new Color(r, g, b);
         }
